refactor(animator): machine and manager tools
